Support Ecto 2.x
Update to support Ecto 2.x. We may still move Ecto support into a separate library, however for now this works for both Ecto 1.x and 2.x.

This commit also includes some minor documentation updates. The documentation still needs plenty of help but I did not want to hold up the Ecto 2.x support for this task.

This also includes a CHANGELOG entry in preparation for release of a new version due to an API update and soft deprecation.