Support loading ecto conditionally
Including graphql_relay in projects which do not use ecto causes problems,
because the ecto dependency here is only required for test, and is not installed
by Mix by default when including graphql_relay as a dependency. This means
that the application won't boot, as it can't find :ecto on launch.